code logic optimize
